Issue #11163: Enforce file size for StaticVariableName input files
